+++ portal-impl/src/com/liferay/portlet/PortletRequestImpl.java
@@ -53,6 +53,7 @@ import com.liferay.portal.security.lang.DoPrivilegedUtil;
 import com.liferay.portal.servlet.NamespaceServletRequest;
 import com.liferay.portal.servlet.SharedSessionServletRequest;
 import com.liferay.portal.util.PropsValues;
+import com.liferay.portlet.internal.PortalContextImpl;
 import com.liferay.portlet.portletconfiguration.util.PublicRenderParameterConfiguration;
 
 import java.security.Principal;
